fix: uspójnić config server i swagger ai-advisor
- przełączenie config-server na backend native dla lokalnego repo konfiguracyjnego

- naprawa endpointów Swagger/OpenAPI w ai-advisor-service

- BigDecimal dla confidence w ai-advisor-service

- poprawa healthchecków PostgreSQL w docker-compose

+++ backend/ai-advisor-service/src/main/java/com/fininsight/advisor/dto/RecommendationResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -15,6 +16,6 @@ import java.util.List;
 public class RecommendationResponse {
     
     private List<String> recommendations;
-    private Double confidence;
+    private BigDecimal confidence;
     private LocalDateTime timestamp;
 }
